Adds DataProvider, adds `ready` method on Application
The use of data sources and data providers allows for a better abstraction and separation between modules that are "feature"-related and modules which only purpose is to provide data to the application. Using the DataProvider abstraction one can create providers for any domain of the Redux State, and through the generated function connect any component to any part of the state without explicitly having to import ActionCreators or creating the same selectors over different modules. This is still very much a WIP.